Optimizations: segmentation into size classes, recycling of freed blocks and alignment of memory blocks + Separation into two folders: “myAllocatorBasic” and “myAllcocator” + Consequential modifications to benchmark, test and CMakeList programs.
The benchmark results after optimizations are :
- malloc/free: 10000 allocations and deallocations in 11630 µs (859845 operations/sec)
- my_malloc/my_free: 10000 allocations and deallocations in 63204 µs (158218 operations/sec)
- my_malloc_basic/my_free_basic: 10000 allocations and deallocations in 24066 µs (415524 operations/sec)
No statistical analysis (average...) has been made, this is just an overview. We can see that the first optimizations didn't directly improve allocation execution speed, due to the fact that adding and managing linked lists adds complexity and time, and that there's a management overhead to maintain the structure (such as periodic cleaning). However, calls to mmap or munmap are reduced (objective of previous commit).
Other forms of benchmarking should also be tested, as for the moment it only deals with the case where allocators allocate several times and then deallocate all at once.

The files have been separated to enable benchmarking and comparison between the initial method (without optimization) and the current method (with optimization).